feat: Implement Leibniz Pi approximation with error calculation
This commit introduces a program that approximates the value of Pi using the Leibniz formula.

The program prompts the user for a number of iterations to determine the accuracy of the calculation. It then iteratively computes the sum.

A helper method, `percentageToPI`, has been added to calculate the percent error of the approximation when compared to the value of `Math.PI`. The final output is formatted to clearly display both the calculated approximation and its percent error.